Splitted tests such that they follow the file structure within src folder.
